+++ src/my_game_list/__init__.py
@@ -1,3 +1,3 @@
 """Main __init__, contains the application version number."""
 
-__version__ = (4, 19, 1)
+__version__ = (4, 20, 0)